+++ packages/blackbox_core/lib/blackbox.dart
@@ -11,11 +11,11 @@ import 'package:meta/meta.dart';
 part 'src/box.dart';
 part 'src/cancel.dart';
 part 'src/connector.dart';
+part 'src/flow_box.dart';
 part 'src/lazy.dart';
 part 'src/output.dart';
 part 'src/persistent.dart';
 part 'src/pipeline.dart';
 part 'src/runtime.dart';
-part 'src/state_observer.dart';
 
 part 'src/test.dart';