+++ packages/blackbox_core/CHANGELOG.md
@@ -1,5 +1,10 @@
-## 0.0.5
-- 
+## 0.0.6
+- Renamed:
+  - StateObserver -> FlowBox
+- Breaking:
+  - `FlowBox<S>` now requires `S extends FlowState`
+- Changed:
+  - FlowBox is now a sync box without input (`Box<O>`)
 ## 0.0.4
 - Renamed:
   - Graph -> Connector